fix integer input to be parsed as double, added q for quit
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@8009 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ misc/equate/src/lex.yy.c
@@ -619,9 +619,9 @@ case 1:
 YY_RULE_SETUP
 #line 15 "calc.l"
 {
-            printf( "An integer: %s (%d)\n", yytext,
-                    atoi( yytext ) );
-            yylval.val= atoi( yytext );
+            printf( "An integer: %s (%g)\n", yytext,
+                    atof( yytext ) );
+            yylval.val= atof( yytext );
             return NUM;
             }
 	YY_BREAK
